Support numpy 2.0
Changes trading index to work exclusively with `int64` dtype.

+++ exchange_calendars/utils/pandas_utils.py
@@ -118,7 +118,7 @@ def longest_run(ser: pd.Series) -> pd.Index:
     ...     | ((ser >= 55) & (ser < 61))
     ... )
     >>> longest_run(bv)
-    Index([30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40], dtype='int32')
+    Index([30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40], dtype='int64')
     >>> pd.testing.assert_index_equal(longest_run(bv), ser.index[30:41])
     """
     # group Trues by only adding to sum when value False.
